refactor: repository 패키지의 utility 패키지명 변경
