+++ src/main/java/chess/repository/dao/TurnDao.java
@@ -2,10 +2,10 @@ package chess.repository.dao;
 
 import chess.model.board.Turn;
 import chess.model.piece.Side;
-import chess.repository.utility.MySqlConnector;
-import chess.repository.utility.ParameterBinder;
-import chess.repository.utility.ResultSetMapper;
-import chess.repository.utility.StatementExecutor;
+import chess.repository.util.MySqlConnector;
+import chess.repository.util.ParameterBinder;
+import chess.repository.util.ResultSetMapper;
+import chess.repository.util.StatementExecutor;
 
 import java.util.Optional;
