+++ src/test/java/chess/repository/dao/PieceDaoTest.java
@@ -9,7 +9,7 @@ import chess.model.position.File;
 import chess.model.position.Position;
 import chess.model.position.Rank;
 import chess.repository.DataBaseCleaner;
-import chess.repository.utility.MySqlConnector;
+import chess.repository.util.MySqlConnector;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
