+++ src/test/java/chess/service/ChessGameServiceTest.java
@@ -14,7 +14,7 @@ import chess.repository.dao.ChessBoardDao;
 import chess.repository.dao.PieceDao;
 import chess.repository.dao.TurnDao;
 import chess.repository.dto.GameResultDto;
-import chess.repository.utility.MySqlConnector;
+import chess.repository.util.MySqlConnector;
 import chess.view.input.MoveArguments;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
